requirements.txt
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
cryptography==2.5
cssselect==1.0.3
enum34==1.1.6
idna==2.8
ipaddress==1.0.22
lxml==4.3.1
Pillow
pycparser==2.19
pyOpenSSL==19.0.0
pyquery==1.4.0
PySocks==1.6.8
requests==2.21.0
selenium==3.141.0
six==1.12.0
urllib3==1.24.1